requirements.txt
Flask
python-dotenv
supabase
google-genai
